work on agent config, bootstrap in auto provision
