+++ agent/agent.go
@@ -5,10 +5,9 @@
 package agent
 
 import (
-	"crypto/tls"
 	"errors"
-	"fmt"
 	"github.com/eclipse/paho.mqtt.golang"
+	"github.com/jmoiron/sqlx"
 	"github.com/ns1labs/orb/agent/backend"
 	"github.com/ns1labs/orb/agent/config"
 	"github.com/ns1labs/orb/agent/policies"
@@ -28,6 +27,7 @@ type orbAgent struct {
 	logger   *zap.Logger
 	config   config.Config
 	client   mqtt.Client
+	db       *sqlx.DB
 	backends map[string]backend.Backend
 
 	hbTicker *time.Ticker
@@ -49,91 +49,21 @@ type orbAgent struct {
 var _ Agent = (*orbAgent)(nil)
 
 func New(logger *zap.Logger, c config.Config) (Agent, error) {
-	pm, err := policies.New(logger, c)
+	var dbFile string
+	if _, ok := c.OrbAgent.DB["file"]; !ok {
+		dbFile = "orb-agent.db"
+	} else {
+		dbFile = c.OrbAgent.DB["file"]
+	}
+	db, err := sqlx.Connect("sqlite3", dbFile)
+	if err != nil {
+		return nil, err
+	}
+	pm, err := policies.New(logger, c, db)
 	if err != nil {
 		return nil, err
 	}
 	return &orbAgent{logger: logger, config: c, policyManager: pm}, nil
-}
-
-func (a *orbAgent) connect() (mqtt.Client, error) {
-
-	opts := mqtt.NewClientOptions().AddBroker(a.config.OrbAgent.MQTT["address"]).SetClientID(a.config.OrbAgent.MQTT["id"])
-	opts.SetUsername(a.config.OrbAgent.MQTT["id"])
-	opts.SetPassword(a.config.OrbAgent.MQTT["key"])
-	opts.SetKeepAlive(2 * time.Second)
-	opts.SetDefaultPublishHandler(func(client mqtt.Client, message mqtt.Message) {
-		a.logger.Info("message on unknown channel, ignoring", zap.String("topic", message.Topic()), zap.ByteString("payload", message.Payload()))
-	})
-	opts.SetPingTimeout(1 * time.Second)
-
-	if !a.config.OrbAgent.TLS.Verify {
-		opts.TLSConfig = &tls.Config{InsecureSkipVerify: true}
-	}
-
-	c := mqtt.NewClient(opts)
-	if token := c.Connect(); token.Wait() && token.Error() != nil {
-		return nil, token.Error()
-	}
-
-	return c, nil
-}
-
-func (a *orbAgent) nameAgentRPCTopics() {
-
-	base := fmt.Sprintf("channels/%s/messages", a.config.OrbAgent.MQTT["channel_id"])
-	a.rpcToCoreTopic = fmt.Sprintf("%s/%s", base, fleet.RPCToCoreTopic)
-	a.rpcFromCoreTopic = fmt.Sprintf("%s/%s", base, fleet.RPCFromCoreTopic)
-	a.capabilitiesTopic = fmt.Sprintf("%s/%s", base, fleet.CapabilitiesTopic)
-	a.heartbeatsTopic = fmt.Sprintf("%s/%s", base, fleet.HeartbeatsTopic)
-	a.logTopic = fmt.Sprintf("%s/%s", base, fleet.LogTopic)
-
-}
-
-func (a *orbAgent) unsubscribeGroupChannels() {
-	for _, channel := range a.groupChannels {
-		if token := a.client.Unsubscribe(channel); token.Wait() && token.Error() != nil {
-			a.logger.Warn("failed to unsubscribe to group channel", zap.String("topic", channel), zap.Error(token.Error()))
-		}
-	}
-}
-
-func (a *orbAgent) startComms() error {
-	var err error
-	a.client, err = a.connect()
-	if err != nil {
-		a.logger.Error("connection failed", zap.Error(err))
-		return err
-	}
-
-	a.nameAgentRPCTopics()
-
-	if token := a.client.Subscribe(a.rpcFromCoreTopic, 1, a.handleRPCFromCore); token.Wait() && token.Error() != nil {
-		a.logger.Error("failed to subscribe to RPC topic", zap.String("topic", a.rpcFromCoreTopic), zap.Error(token.Error()))
-		return token.Error()
-	}
-
-	err = a.sendCapabilities()
-	if err != nil {
-		a.logger.Error("failed to send agent capabilities", zap.Error(err))
-		return err
-	}
-
-	err = a.sendGroupMembershipReq()
-	if err != nil {
-		a.logger.Error("failed to send group membership request", zap.Error(err))
-	}
-
-	err = a.sendAgentPoliciesReq()
-	if err != nil {
-		a.logger.Error("failed to send agent policies request", zap.Error(err))
-	}
-
-	a.hbTicker = time.NewTicker(HeartbeatFreq)
-	a.hbDone = make(chan bool)
-	go a.sendHeartbeats()
-
-	return nil
 }
 
 func (a *orbAgent) startBackends() error {
@@ -171,42 +101,20 @@ func (a *orbAgent) Start() error {
 	mqtt.CRITICAL = &agentLoggerCritical{a: a}
 	mqtt.ERROR = &agentLoggerError{a: a}
 
+	cloudConfig, err := GetCloudConfig(a.config, a.db)
+	if err != nil {
+		return err
+	}
+
 	if err := a.startBackends(); err != nil {
 		return err
 	}
 
-	if err := a.startComms(); err != nil {
+	if err := a.startComms(cloudConfig); err != nil {
 		return err
 	}
 
 	return nil
-}
-
-func (a *orbAgent) subscribeGroupChannels(groups []fleet.GroupMembershipData) []string {
-	var successList []string
-	for _, groupData := range groups {
-
-		base := fmt.Sprintf("channels/%s/messages", groupData.ChannelID)
-		rpcFromCoreTopic := fmt.Sprintf("%s/%s", base, fleet.RPCFromCoreTopic)
-
-		token := a.client.Subscribe(rpcFromCoreTopic, 1, a.handleGroupRPCFromCore)
-		if token.Error() != nil {
-			a.logger.Error("failed to subscribe to group channel/topic", zap.String("topic", rpcFromCoreTopic), zap.Error(token.Error()))
-			continue
-		}
-		ok := token.WaitTimeout(time.Second * 5)
-		if ok && token.Error() != nil {
-			a.logger.Error("failed to subscribe to group channel/topic", zap.String("topic", rpcFromCoreTopic), zap.Error(token.Error()))
-			continue
-		}
-		if !ok {
-			a.logger.Error("failed to subscribe to group channel/topic: time out", zap.String("topic", rpcFromCoreTopic))
-			continue
-		}
-		a.logger.Info("completed RPC subscription to group", zap.String("name", groupData.Name), zap.String("topic", rpcFromCoreTopic))
-		successList = append(successList, groupData.ChannelID)
-	}
-	return successList
 }
 
 func (a *orbAgent) Stop() {
